chore: moved unit tests next to implementation
