+++ src/card/interface/gql.test.js
@@ -1,0 +1,10 @@
+/* @flow */
+import { describe, it, expect } from "vitest";
+
+import { resetGQLErrors, emitGqlErrors } from "./gql";
+
+describe("card/interface/gql", () => {
+  it("should be true", () => {
+    expect(true).toBe(true);
+  });
+});
