--- tests/card/interface/gql.test.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest';
-
-import { resetGQLErrors, emitGqlErrors } from '../../../src/card/interface';
-
-describe('card/interface/gql', () => {
-  it('should be true', () => {
-    expect(true).toBe(true);
-  });
-})
